Schema setup and model and controller creation
Set up individual tables in /database/migrations
Created models and controllers for each table
Simple test for /api/assets

+++ adonis/database/migrations/1531607105447_price_schema.js
@@ -3,24 +3,27 @@
 const Schema = use("Schema");
 
 class PriceSchema extends Schema {
-  up() {
-    this.create("prices", table => {
-      table.increments("price_id");
-      table.integer("exchange_pair_id");
-      table.float("price");
-      table.float("bid");
-      table.float("ask");
-      table.float("base_volume");
-      table.float("quote_volume");
-      table.timestamps();
-    });
-    this.alter("pairs", table => {
-      table
-        .integer("price_exchchange_pair_fkey")
-        .references("exchange_pair_id")
-        .inTable("exchange_pairs")
-        .onDelete("cascade");
-    });
+  async up() {
+    const exists = await this.hasTable("exchange_pair_id");
+    if (exists) {
+      this.create("prices", table => {
+        table.increments("price_id");
+        table.integer("exchange_pair_id");
+        table.float("price").notNullable();
+        table.float("bid");
+        table.float("ask");
+        table.float("base_volume");
+        table.float("quote_volume");
+        table.timestamps();
+      });
+      this.alter("prices", table => {
+        table
+          .integer("price_exchchange_pair_fkey")
+          .references("exchange_pair_id")
+          .inTable("exchange_pairs")
+          .onDelete("cascade");
+      });
+    }
   }
 
   down() {
