+++ adonis/database/migrations/1531608013508_exchange_pair_schema.js
@@ -3,30 +3,35 @@
 const Schema = use("Schema");
 
 class ExchangePairSchema extends Schema {
-  up() {
-    this.create("exchange_pairs", table => {
-      table.increments("exchange_pair_id");
-      table
-        .integer("exchange_id")
-        .references("exchange_id")
-        .inTable("exchanges")
-        .onDelete("cascade")
-        .unique();
-      table
-        .integer("pair_id")
-        .references("pair_id")
-        .inTable("pairs")
-        .onDelete("cascade")
-        .unique();
-      table
-        .integer("last_price")
-        .references("price_id")
-        .inTable("prices")
-        .onDelete("set null");
-      table.boolean("active");
-      table.json("mkt_precision");
-      table.timestamps();
-    });
+  async up() {
+    const exist = await this.hasTable("prices");
+    if (exist) {
+      this.create("exchange_pairs", table => {
+        table.increments("exchange_pair_id");
+        table
+          .integer("exchange_id")
+          .references("exchange_id")
+          .inTable("exchanges")
+          .onDelete("cascade")
+          .unique()
+          .notNullable();
+        table
+          .integer("pair_id")
+          .references("pair_id")
+          .inTable("pairs")
+          .onDelete("cascade")
+          .unique()
+          .notNullable();
+        table
+          .integer("last_price")
+          .references("price_id")
+          .inTable("prices")
+          .onDelete("set null");
+        table.boolean("active");
+        table.json("mkt_precision");
+        table.timestamps();
+      });
+    }
   }
 
   down() {
